requirements.txt
opencv-python
numpy
pandas
Pillow
PytorchWildlife
supervision
torch
tqdm
transformers
